+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 * Fixed crash when depositing materials
 * /kingdomkeys material command when setting or taking materials.
 * A couple of translations.
+* Drive form command menu being too wide.
+* Limit attacks having no cooldown.